requirements.txt
pandas
SQLAlchemy
psycopg[binary]
python-dotenv
